Room side menu

+++ src/api/api.js
@@ -1,31 +1,67 @@
 import axios from 'axios';
+import jwt_decode from 'jwt-decode';
 
 export const backendUrl = 'http://localhost:5000';
 
 class ApiService {
   httpGET(url) {
-    return axios.get(backendUrl + url, this.authorize())
-      .then(this.mapResponse);
+    if (this.validateToken()) {
+      return axios.get(backendUrl + url, this.authorize())
+        .then(this.mapResponse);
+    } else {
+      return this.ifExpired()
+    }
+  }
+
+  httpGETAvatar(url) {
+    if (this.validateToken()) {
+      return axios.get(backendUrl + url, this.authorizeWithBlob())
+        .then(this.mapResponse);
+    } else {
+      return this.ifExpired()
+    }
   }
 
   httpPOST(url, data = {}) {
+    if (this.validateToken()) {
+      return axios.post(backendUrl + url, data, this.authorize())
+        .then(this.mapResponse);
+    } else {
+      return this.ifExpired()
+    }
+  }
+
+  httpAuthPOST(url, data = {}) {
     return axios.post(backendUrl + url, data, this.authorize())
       .then(this.mapResponse);
   }
 
   httpDELETE(url) {
-    return axios.delete(backendUrl + url, this.authorize())
-      .then(this.mapResponse);
+    if (this.validateToken()) {
+      return axios.delete(backendUrl + url, this.authorize())
+        .then(this.mapResponse);
+    } else {
+      return this.ifExpired()
+    }
   }
 
   httpPUT(url, data = {}) {
-    return axios.put(backendUrl + url, data, this.authorize())
-      .then(this.mapResponse);
+    if (this.validateToken()) {
+      return axios.put(backendUrl + url, data, this.authorize())
+        .then(this.mapResponse);
+    } else {
+      return this.ifExpired()
+    }
   }
 
-  httpPATCH(url, data = {}){
-    return axios.patch(backendUrl + url, data, this.authorize())
-    .then(this.mapResponse);
+
+  httpPATCH(url, data = {}) {
+    if (this.validateToken()) {
+      return axios.patch(backendUrl + url, data, this.authorize())
+        .then(this.mapResponse);
+    } else {
+      return this.ifExpired()
+    }
   }
 
   mapResponse = (response) => response.data
@@ -33,9 +69,47 @@ class ApiService {
   authorize = () => {
     const token = JSON.parse(localStorage.getItem('token'));
     if (token) {
-      return { headers: { Authorization: 'Bearer ' + token } };
+      return {
+        headers: {
+          Authorization: 'Bearer ' + token,
+        }
+      };
     }
     return null;
+  }
+
+  authorizeWithBlob = () => {
+    const token = JSON.parse(localStorage.getItem('token'));
+    if (token) {
+      return {
+        responseType: 'blob',
+        headers: {
+          Authorization: 'Bearer ' + token,
+        }
+      };
+    }
+    return null;
+  }
+
+  validateToken = () => {
+    const decodedToken = jwt_decode(localStorage.getItem('token'))
+    const currentDate = new Date();
+    if (decodedToken.exp * 1000 < currentDate.getTime()) {
+      console.log("Token expired.");
+      localStorage.removeItem('token')
+      localStorage.setItem('tokenExpired', true);
+      return false;
+    } else {
+      console.log("Valid token");
+      return true;
+    }
+  }
+
+  ifExpired = () => {
+    window.location.reload(false);
+    return new Promise((resolve) => {
+      resolve(console.log("Token expired, redirect to default page"));
+    })
   }
 }
